+++ libs/admin/src/wagoe/admin/shell/http/handlers/detail.clj
@@ -102,7 +102,7 @@
               permissions (permissions/get-entity-permissions user entity-name entity-config)]
           (support/html-response request
                                  (admin-ui/admin-layout
-                                  (admin-ui/entity-detail-page entity-name entity-config nil {} permissions {})
+                                  (admin-ui/entity-detail-page entity-name entity-config nil {} permissions {:display (support/display-options config request)})
                                   {:user user
                                    :current-entity entity-name
                                    :entities entities